requirements.txt
annotated-types==0.7.0
anyio==4.9.0
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
docutils==0.21.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.10.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
openai==1.79.0
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
sniffio==1.3.1
tqdm==4.67.1
twine==6.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
